requirements.txt
Click==7.0
connexion[swagger-ui]==2.3.0
Faker==2.0.1
Flask==1.1.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
flask-marshmallow==0.10.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
pytest==5.1.1
python-dotenv==0.10.3
WTForms==2.2.1
